Themes

Apply a pre-built or custom coded theme to a chart

Usage

ec.theme(wt, name = "custom", code = NULL)

Arguments

wt

Required echarty widget as returned by ec.init

name

Name of existing theme file (without extension), or name of custom theme defined in code.

code

Custom theme as JSON formatted string, default NULL.

Value

An echarty widget.

Details

Just a few built-in themes are included in folder inst/themes.
Their names are dark, gray, jazz, dark-mushroom and macarons.
The entire ECharts theme collection could be found here and files copied if needed.
To create custom themes or view predefined ones, visit this site.

Examples

mtcars |> ec.init() |> ec.theme('dark-mushroom')

cars |> ec.init() |> ec.theme('mine', code=
  '{"color": ["green","#eeaa33"], 
    "backgroundColor": "lemonchiffon"}')